Updated Chase- Fixed OnTrigger Function
I fixed a typo so Collisions (OnTrigger)  now register between the square and the player.

I added an extra Box Collider to the player and the Square. - You need two. One for registering collions (the one with Is Trigger enabled) and the other to keep them from warping into each other.